+++ src/client/css/custom/swap.scss
@@ -449,7 +449,7 @@ $swap-widget-anim-timing: 300ms;
     }
   }
 
-  .warning-view {
+  .custom-token-warning {
     width: 100%;
     padding: 1rem;
     border-radius: 16px;
